requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-colorfield==0.4.1
django-crispy-forms==1.11.2
django-filter==2.4.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
gunicorn==20.1.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==2.0.0
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0